add protect on ex01

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: achansar <achansar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/13 15:19:37 by achansar          #+#    #+#             */
-/*   Updated: 2023/07/26 17:14:56 by achansar         ###   ########.fr       */
+/*   Updated: 2023/07/28 13:36:05 by achansar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ int main(int argc, char **argv)
     if (argc == 1)
         std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
     else {
-        for (int j = 1; j < argc; j++) {
-            for (int i = 0; i < strlen(argv[j]); i++)
+        for (size_t j = 1; j < argc; j++) {
+            for (size_t i = 0; i < strlen(argv[j]); i++)
                 std::cout << (char)toupper(argv[j][i]);
         }
         std::cout << std::endl;
